requirements.txt
gradio==5.42.0
manim==0.19.0
Pillow==11.3.0
svgpathtools==1.7.1
